Move useInternalValue to hooks directory inside components
